Renamed config module to conf
